Get code working again after merge.  Had to reconcile the move of the gradebook classes to their own package.
